ai-image-captioner/requirements.txt
streamlit==1.32.0
transformers==4.41.2
Pillow==10.0.0
accelerate==0.22.0

